impemented ensamble. random forest. 82.1. and analysis of low score
lower score than before as earlier, max_groups_in_categorical_data was not used. earlier all names and some ticket columns were checked for groups. all groups in a column were kept and non groups were replaced as nan. However, this can cause the  algorithms to remember peoples names (ie unique identifiers) causing over fitting. with the introduction of new variable in if statement at line 617, we check if there are too many groups. this makes name group get converted to hash. in future we can try a different function for char and a different for numbers. it is not clear if md5 is useful.